第一问完成

+++ Ccode/red_firm/project/code/easy_ui.h
@@ -79,7 +79,7 @@ extern float GetBatteryVoltage();
 #define EasyUIGetBatVoltage()                                   (GetBatteryVoltage())
 #define EasyUIDelay_ms(time)                                    (system_delay_ms(time))
 
-typedef     double      paramType;
+typedef     float      paramType;
 
 typedef enum
 {
